Rename _func_w
