+++ lib/contractions/pipi_scattering_func_w.h
